Feat: 로그인/회원가입 개선 및 설문 템플릿 구조 분리 (#14)
